BCD_Models/requirements.txt
Django==5.1.3
djangorestframework==3.15.1
firebase-admin==6.5.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.4.2
tensorflow==2.19.0
Pillow==10.4.0
cloudinary==1.41.0
python-dotenv==1.0.1
gunicorn==22.0.0
whitenoise==6.7.0